Business pages implementation

+++ CMS/Controllers/BusinessPagesController.cs
@@ -18,28 +18,38 @@ namespace CMS.Controllers
         // GET: BusinessPages
         public ActionResult Index()
         {
-            return View(db.BusinessPages.ToList());
+            if ((Session["UserID"] != null)) {
+                return View(db.BusinessPages.ToList());
+            } else {
+                return RedirectToAction("Login", "Home");
+            }
         }
 
         // GET: BusinessPages/Details/5
         public ActionResult Details(long? id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            BusinessPage businessPage = db.BusinessPages.Find(id);
-            if (businessPage == null)
-            {
-                return HttpNotFound();
-            }
-            return View(businessPage);
+            if ((Session["UserID"] != null)) {
+                if (id == null) {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                BusinessPage businessPage = db.BusinessPages.Find(id);
+                if (businessPage == null) {
+                    return HttpNotFound();
+                }
+                return View(businessPage);
+            } else {
+                return RedirectToAction("Login", "Home");
+            }            
         }
 
         // GET: BusinessPages/Create
         public ActionResult Create()
         {
-            return View();
+            if ((Session["UserID"] != null)) {
+                return View();
+            } else {
+                return RedirectToAction("Login", "Home");
+            }
         }
 
         // POST: BusinessPages/Create
@@ -49,31 +59,41 @@ namespace CMS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "PageName,PageNameUrl,PageContent")] BusinessPage businessPage)
         {
-            if (ModelState.IsValid)
-            {
-                businessPage.CreatedDate = DateTime.Now;
-                businessPage.UpdateDate = DateTime.Now;
-                db.BusinessPages.Add(businessPage);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+            if ((Session["UserID"] != null)) {
+                if (ModelState.IsValid) {
+                    if (!db.BusinessPages.Any(x => x.PageNameUrl == businessPage.PageNameUrl)) {
+                        businessPage.CreatedDate = DateTime.Now;
+                        businessPage.UpdateDate = DateTime.Now;
+                        db.BusinessPages.Add(businessPage);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    } else {
+                        ModelState.AddModelError("PageNameUrl", "PageNameUrl Already Exist.");
+                        return View(businessPage);
+                    }
+                }
 
-            return View(businessPage);
+                return View(businessPage);
+            } else {
+                return RedirectToAction("Login", "Home");
+            }           
         }
 
         // GET: BusinessPages/Edit/5
         public ActionResult Edit(long? id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            BusinessPage businessPage = db.BusinessPages.Find(id);
-            if (businessPage == null)
-            {
-                return HttpNotFound();
-            }
-            return View(businessPage);
+            if ((Session["UserID"] != null)) {
+                if (id == null) {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                BusinessPage businessPage = db.BusinessPages.Find(id);
+                if (businessPage == null) {
+                    return HttpNotFound();
+                }
+                return View(businessPage);
+            } else {
+                return RedirectToAction("Login", "Home");
+            }           
         }
 
         // POST: BusinessPages/Edit/5
@@ -81,47 +101,60 @@ namespace CMS.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "BusinessPageID,PageName,PageNameUrl,PageContent")] BusinessPage businessPage)
+        public ActionResult Edit([Bind(Include = "BusinessPageID,PageName,PageNameUrl,PageContent,CreatedDate,UpdateDate")] BusinessPage businessPage)
         {
-            if (ModelState.IsValid)
-            {
-                //var currentBusinessPage = db.BusinessPages.FirstOrDefault(p => p.BusinessPageID == businessPage.BusinessPageID);
-                //if (currentBusinessPage == null)
-                //    return HttpNotFound();
-                businessPage.UpdateDate = DateTime.Now;
-                businessPage.CreatedDate = DateTime.Now;// currentBusinessPage.CreatedDate;
-                db.Entry(businessPage).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            return View(businessPage);
+            if ((Session["UserID"] != null)) {
+                if (ModelState.IsValid) {
+                    if (!db.BusinessPages.Any(x => x.PageNameUrl == businessPage.PageNameUrl && x.BusinessPageID != businessPage.BusinessPageID)) {
+                        businessPage.UpdateDate = DateTime.Now;
+                        db.Entry(businessPage).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    } else {
+                        ModelState.AddModelError("PageNameUrl", "PageNameUrl Already Exist.");
+                        return View(businessPage);
+                    }
+                }
+                return View(businessPage);
+            } else {
+                return RedirectToAction("Login", "Home");
+            }            
         }
 
         // GET: BusinessPages/Delete/5
         public ActionResult Delete(long? id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            BusinessPage businessPage = db.BusinessPages.Find(id);
-            if (businessPage == null)
-            {
-                return HttpNotFound();
-            }
-            return View(businessPage);
+            if ((Session["UserID"] != null)) {
+                if (id == null) {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                BusinessPage businessPage = db.BusinessPages.Find(id);
+                if (businessPage == null) {
+                    return HttpNotFound();
+                }
+                return View(businessPage);
+            } else {
+                return RedirectToAction("Login", "Home");
+            }            
         }
 
         // POST: BusinessPages/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(long id)
-        {
-            BusinessPage businessPage = db.BusinessPages.Find(id);
-            db.BusinessPages.Remove(businessPage);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+        public ActionResult DeleteConfirmed(long id) {
+            if ((Session["UserID"] != null)) {
+                BusinessPage businessPage = db.BusinessPages.Find(id);
+                db.BusinessPages.Remove(businessPage);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            } else {
+                return RedirectToAction("Login", "Home");
+            }
         }
+
+        //public JsonResult IsPageNameUrlUnique(string PageNameUrl) {
+        //    return Json(!db.BusinessPages.Any(x => x.PageNameUrl == PageNameUrl), JsonRequestBehavior.AllowGet);
+        //}
 
         protected override void Dispose(bool disposing)
         {
